+++ src/pykmextract/bridge/__init__.py
@@ -1,5 +1,0 @@
-"""Bridge integrations."""
-
-from .pyheor import PyHEORBridge
-
-__all__ = ["PyHEORBridge"]